HTML output method
New file html.lisp with a SAX handler that sits in front of HAX/SAX
sinks and implements the XSLT-specific details of HTML serialization.

New slot `encoding' on the output specification.

Use a `method' slot on the output specification instead of the old
stylesheet-html-output-p.

In the tests, ignore the `compare' attribute of the test case, it's not
set correctly.  Instead, parse the stylesheet to find the actual method
value.

Compare HTML output using the parse tree from Closure HTML.
Normalize whitespace (mostly) according to HTML's rules before comparing.