Fixed result tree fragments.
Result tree fragments are not node sets, so don't wrap them in node set
instances.

Declare result tree fragments as XPath nodes, so that the string-value
protocol method will be used on them.

Resignal XPath errors, so that an XSLT error will be signalled when a user
tries to use a result tree fragment where only a node set is allowed.