Store variable values in arrays.
Instead of a list of scopes, there are now only two scopes:
  - the array *global-variable-values*, bound once
  - and the array *lexical-variable-values*, bound for each template call

Each xsl:variable gets its own index in that array.  The total number of
lexical variables is stored for each template, so that an array of the
right size can be created when the call is made.

Together with the previous patch moving the logic for duplicate variable
checks into the parser, this commit removed the need for scoped-progn.

During compilation, we keep track of the variable declarations in
  - *global-variable-declarations*
  - *lexical-variable-declarations*
  - as well as the initial-global-variable-thunks as before
Currently, the first two variables hold an array rather than a list, which
is probably a mistake in the case of *lexical-variable-declarations*.  See
the kludge with FILL in the implementation of LET.  To be fixed later.