Dubious namespace alias workarounds:
Sometimes the tests want excluded namespace URIs in the output.  All
cases of this issue that I can find involve the target URIs of namespace
aliases.

XSLT 1.0 is silent about this, XSLT 2.0 more explicit (but they rename
prefixes, so their behaviour doesn't match the test suite either).

We need to fix this, because the usual case of stylesheet-generating
stylesheets is affected by it.

This kludge can't really be the answer though, and actually breaks one
other test case.  Sigh.  I'm still checking it in because it improves our
overall test score.